pyproject.toml
[project]
name = "opcut"
version = "0.4.15"
description = "Cutting stock problem optimizer"
readme = "README.rst"
requires-python = ">=3.10"
license = {text = "GPLv3"}
dependencies = [
    "aiohttp ~=3.10.11",
    "hat-aio ~=0.7.11",
    "hat-json ~=0.6.1",
    "pycairo ~=1.27.0",
]

[project.scripts]
opcut = "opcut.main:main"

[project.urls]
Homepage = "https://opcut.kopic.xyz"
Repository = "https://github.com/bozokopic/opcut.git"

[project.optional-dependencies]
dev = ["hat-doit ~=0.16.1"]

[build-system]
requires = ["hat-doit ~=0.16.1"]
build-backend = "hat.doit.pep517"

[tool.hat-doit]
build_wheel_task = "wheel"
